backend/requirements.txt
django=3.2
djangorestframework=3.12.4